Move API endpoints into separate version directories
