+++ api/data_workspace/v1/tests/test_licence_views.py
@@ -1,12 +1,6 @@
-import mohawk
-
-from django.conf import settings
-from django.test import override_settings
 from django.urls import reverse
 from rest_framework import status
-from urllib import parse
 
-from api.core.requests import get_hawk_sender
 from api.applications.tests.factories import GoodOnApplicationFactory
 from api.cases.tests.factories import FinalAdviceFactory
 from api.cases.enums import AdviceType
@@ -18,9 +12,6 @@ from test_helpers.clients import DataTestClient
 class DataWorkspaceTests(DataTestClient):
     def setUp(self):
         super().setUp()
-        test_host = "http://testserver"
-        self.licences = parse.urljoin(test_host, reverse("data_workspace:dw-licences-only-list"))
-        self.ogl_list = parse.urljoin(test_host, reverse("data_workspace:dw-ogl-only-list"))
         # Set up fixtures for testing.
         case = self.create_standard_application_case(self.organisation)
         good = GoodFactory(
@@ -28,9 +19,7 @@ class DataWorkspaceTests(DataTestClient):
             is_good_controlled=True,
             control_list_entries=["ML21"],
         )
-        good_advice = FinalAdviceFactory(
-            user=self.gov_user, team=self.team, case=case, good=good, type=AdviceType.APPROVE
-        )
+        FinalAdviceFactory(user=self.gov_user, team=self.team, case=case, good=good, type=AdviceType.APPROVE)
         GoodOnLicenceFactory(
             good=GoodOnApplicationFactory(application=case, good=good),
             licence=StandardLicenceFactory(case=case),
@@ -38,36 +27,8 @@ class DataWorkspaceTests(DataTestClient):
             value=1,
         )
 
-    @override_settings(HAWK_AUTHENTICATION_ENABLED=True)
-    def test_dw_view_licences(self):
-        sender = get_hawk_sender("GET", self.licences, None, settings.HAWK_LITE_DATA_WORKSPACE_CREDENTIALS)
-        self.client.credentials(HTTP_HAWK_AUTHENTICATION=sender.request_header, CONTENT_TYPE="application/json")
-        response = self.client.get(self.licences)
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-    @override_settings(HAWK_AUTHENTICATION_ENABLED=True)
-    def test_dw_view_licences_fail_incorrect_hawk_key(self):
-        sender = get_hawk_sender("GET", self.licences, None, "internal-frontend")
-        self.client.credentials(HTTP_HAWK_AUTHENTICATION=sender.request_header, CONTENT_TYPE="application/json")
-        with self.assertRaises(mohawk.exc.HawkFail):
-            self.client.get(self.licences)
-
-    @override_settings(HAWK_AUTHENTICATION_ENABLED=True)
-    def test_dw_view_ogl_types(self):
-        sender = get_hawk_sender("GET", self.ogl_list, None, settings.HAWK_LITE_DATA_WORKSPACE_CREDENTIALS)
-        self.client.credentials(HTTP_HAWK_AUTHENTICATION=sender.request_header, CONTENT_TYPE="application/json")
-        response = self.client.get(self.ogl_list)
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-    @override_settings(HAWK_AUTHENTICATION_ENABLED=True)
-    def test_dw_view_ogl_fail_incorrect_hawk_key(self):
-        sender = get_hawk_sender("GET", self.ogl_list, None, "internal-frontend")
-        self.client.credentials(HTTP_HAWK_AUTHENTICATION=sender.request_header, CONTENT_TYPE="application/json")
-        with self.assertRaises(mohawk.exc.HawkFail):
-            self.client.get(self.ogl_list)
-
     def test_good_on_licenses(self):
-        url = reverse("data_workspace:dw-good-on-licences-list")
+        url = reverse("data_workspace:v1:dw-good-on-licences-list")
         expected_fields = (
             "good_on_application_id",
             "usage",
@@ -99,7 +60,7 @@ class DataWorkspaceTests(DataTestClient):
         self.assertEqual(tuple(options.keys()), expected_fields)
 
     def test_licenses(self):
-        url = reverse("data_workspace:dw-licences-list")
+        url = reverse("data_workspace:v1:dw-licences-list")
         expected_fields = ("id", "reference_code", "status", "application", "goods")
         response = self.client.get(url)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
